query 2 done

+++ include/posts.h
@@ -7,18 +7,18 @@
 typedef struct posts* POSTS;
 
 POSTS create_posts();
-gint get_key_id_post(POSTS p);
+long get_key_id_post(POSTS p);
 Date get_data(POSTS p);
-gint get_score(POSTS p);
-gint get_owner_user_id(POSTS p);
+long get_score(POSTS p);
+long get_owner_user_id(POSTS p);
 gchar* get_title(POSTS p);
 gchar* get_body(POSTS p);
-gint get_post_type_id(POSTS p);
-gint get_parent_id(POSTS p);
+long get_post_type_id(POSTS p);
+long get_parent_id(POSTS p);
 gchar* get_tags(POSTS p);
-gint get_answer_count(POSTS p);
-gint get_comment_count(POSTS p);
-gint get_dif_votes(POSTS p);
+long get_answer_count(POSTS p);
+long get_comment_count(POSTS p);
+long get_dif_votes(POSTS p);
 void set_key_id_post(POSTS p, int key_id_post);
 void set_data(POSTS p, Date data);
 void set_score(POSTS p, int score);
